chore(vscode): canvas adapter consumes @cc-wf-studio/mcp factory
- McpServerManager now implements WorkflowIoAdapter and is passed straight
  to createWorkflowMcpServer in the HTTP request handler
- Per-request server identity uses the @cc-wf-studio/mcp defaults so MCP
  clients see the same metadata as before
- planSubAgentFiles + executeSubAgentFileCreation move inline next to the
  manager (canvas-mode-specific auto-create)
- Delete packages/vscode/src/extension/services/mcp-server-tools.ts
- Add @cc-wf-studio/mcp as workspace dep on cc-wf-studio
- Revision crosses the adapter boundary as a string (canvas-side number is
  stringified; expectedRevision is parsed back on the way in)

Phase 3 step 3/6 of @cc-wf-studio/mcp standalone server
